The earlier code did not have proper error handling for URL error, when website is down. RMA was done on one of the days and web scraping failed. To rectify this, an error handling has been implemented. In order to check if this new script works fine, will implement the new scripts with a suffix "2" after all file names. - BTN.Exhange.Rate.Scraping_v2.R - scrapeBOB2.R - scrapeRMA2.R - scrapeTBL2.R
New file scraped will be suffixed with "_new" before the ".csv" file extension.

Error handling codes are implemented in the scrapeBank R scripts and link to these are changed in the main file (_V2.R).

GitHub action for this file to be created with new name daily_new.

Future code revisions to explore:
- same scraping file with different parameters based on Bank being scraped.
- this might make future changes to scripts easier.
- re-organizing code into functions for establishing link, scraping content, checking format consistency and cleaning columns, etc.